+++ src/components/client/PinkMap/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useMapAutoReturn } from './useMapAutoReturn';
